requirements.txt
asgiref==3.3.1
backcall==0.2.0
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
click==7.1.2
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
decorator==4.4.2
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
Flask==1.1.2
Flask-UUID==0.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.7.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytz==2021.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
six @ file:///tmp/build/80754af9/six_1605205327372/work
sqlparse==0.4.1
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==5.0.5
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
uuid==1.30
wcwidth==0.2.5
Werkzeug==1.0.1
